+++ src/formatters/expression.rs
@@ -392,7 +392,9 @@ fn format_token_expression_sequence(
 
     let requires_multiline_expression = shape.take_first_line(&formatted_expression).over_budget()
         || trivia_util::token_contains_trailing_comments(token)
-        || trivia_util::contains_comments(expression);
+        || trivia_util::contains_comments(
+            expression.update_trailing_trivia(FormatTriviaType::Replace(vec![])),
+        ); // Remove trailing trivia (comments) before checking, as they shouldn't have an impact
 
     let token = match requires_multiline_expression {
         // `<token>\n`
@@ -596,10 +598,12 @@ fn format_if_expression(ctx: &Context, if_expression: &IfExpression, shape: Shap
         );
 
         // Put the else on a new line
-        let else_token = else_token.update_leading_trivia(FormatTriviaType::Append(vec![
-            create_newline_trivia(ctx),
-            create_indent_trivia(ctx, hanging_shape),
-        ]));
+        let else_token = trivia_util::prepend_newline_indent(
+            ctx,
+            &else_token,
+            else_token.leading_trivia(),
+            hanging_shape,
+        );
 
         IfExpression::new(condition, expression, else_expression)
             .with_if_token(if_token)