+++ src/formatters/luau.rs
@@ -11,12 +11,12 @@ use crate::{
         table::{create_table_braces, format_multiline_table, format_singleline_table, TableType},
         trivia::{
             strip_leading_trivia, strip_trailing_trivia, strip_trivia, FormatTriviaType,
-            UpdateLeadingTrivia, UpdateTrailingTrivia,
+            UpdateLeadingTrivia, UpdateTrailingTrivia, UpdateTrivia,
         },
         trivia_util::{
             contains_comments, take_type_argument_trailing_comments,
             token_trivia_contains_comments, trivia_is_comment, trivia_is_newline,
-            type_info_trailing_trivia,
+            type_info_leading_trivia, type_info_trailing_trivia,
         },
     },
     shape::Shape,
@@ -336,32 +336,72 @@ pub fn format_type_info(ctx: &Context, type_info: &TypeInfo, shape: Shape) -> Ty
     }
 }
 
-pub fn hang_type_info(ctx: &Context, type_info: TypeInfo, shape: Shape) -> TypeInfo {
-    match type_info {
-        TypeInfo::Union { left, pipe, right } => {
-            let mut pipe_leading_trivia: Vec<_> = pipe
-                .leading_trivia()
+// A clone of [`hang_binop`], except for TypeInfo tokens. TODO: can we merge the two?
+fn hang_type_info_binop(
+    ctx: &Context,
+    binop: TokenReference,
+    shape: Shape,
+    rhs: &TypeInfo,
+) -> TokenReference {
+    // Get the leading comments of a binop, as we need to preserve them
+    // Intersperse a newline and indent trivia between them
+    // iter_intersperse is currently not available, so we need to do something different. Tracking issue: https://github.com/rust-lang/rust/issues/79524
+    let leading_comments = binop
+        .leading_trivia()
+        .filter(|token| trivia_is_comment(token))
+        .flat_map(|x| {
+            vec![
+                create_newline_trivia(ctx),
+                create_indent_trivia(ctx, shape),
+                x.to_owned(),
+            ]
+        })
+        // If there are any comments trailing the BinOp, we need to move them to before the BinOp
+        .chain(
+            binop
+                .trailing_trivia()
                 .filter(|token| trivia_is_comment(token))
-                .flat_map(|trivia| {
-                    // Prepend an indent before the comment, and append a newline after the comments
+                .map(|x| x.to_owned()),
+        )
+        // If there are any leading comments to the RHS expression, we need to move them to before the BinOp
+        .chain(
+            type_info_leading_trivia(rhs)
+                .iter()
+                .filter(|token| trivia_is_comment(token))
+                .flat_map(|x| {
                     vec![
                         create_newline_trivia(ctx),
                         create_indent_trivia(ctx, shape),
-                        trivia.to_owned(),
+                        x.to_owned().to_owned(),
                     ]
-                })
-                .collect();
+                }),
+        )
+        // Create a newline just before the BinOp, and preserve the indentation
+        .chain(std::iter::once(create_newline_trivia(ctx)))
+        .chain(std::iter::once(create_indent_trivia(ctx, shape)))
+        .collect();
 
-            pipe_leading_trivia.push(create_newline_trivia(ctx));
-            pipe_leading_trivia.push(create_indent_trivia(ctx, shape));
+    binop.update_trivia(
+        FormatTriviaType::Replace(leading_comments),
+        FormatTriviaType::Replace(vec![Token::new(TokenType::spaces(1))]),
+    )
+}
 
-            TypeInfo::Union {
-                left,
-                pipe: pipe.update_leading_trivia(FormatTriviaType::Replace(pipe_leading_trivia)),
-                right: Box::new(hang_type_info(ctx, *right, shape)),
-            }
-        }
-        _ => type_info,
+/// Hangs a type info at a pipe operator, then reformats either side with the new shape
+pub fn hang_type_info(ctx: &Context, type_info: &TypeInfo, shape: Shape) -> TypeInfo {
+    const PIPE_LENGTH: usize = 2; // "| "
+
+    match type_info {
+        TypeInfo::Union { left, pipe, right } => TypeInfo::Union {
+            left: Box::new(format_type_info(ctx, left, shape)),
+            pipe: hang_type_info_binop(ctx, pipe.to_owned(), shape, right),
+            right: Box::new(hang_type_info(
+                ctx,
+                &right.update_leading_trivia(FormatTriviaType::Replace(vec![])),
+                shape.reset() + PIPE_LENGTH,
+            )),
+        },
+        other => format_type_info(ctx, other, shape),
     }
 }
 
@@ -552,20 +592,21 @@ fn format_type_declaration(
     };
 
     let mut equal_token = fmt_symbol!(ctx, type_declaration.equal_token(), " = ", shape);
-    let mut type_definition =
-        format_type_info(ctx, type_declaration.type_definition(), shape + 3) // 3 = " = "
-            .update_trailing_trivia(FormatTriviaType::Append(trailing_trivia));
+    let mut type_definition = format_type_info(ctx, type_declaration.type_definition(), shape + 3); // 3 = " = "
 
-    let shape = shape.take_last_line(&strip_trailing_trivia(&type_definition));
-
-    if should_hang_type(type_declaration.type_definition()) || shape.over_budget() {
+    if should_hang_type(type_declaration.type_definition())
+        || shape.test_over_budget(&strip_trailing_trivia(&type_definition))
+    {
         let shape = shape.increment_additional_indent();
         equal_token = equal_token.update_trailing_trivia(FormatTriviaType::Replace(vec![
             create_newline_trivia(ctx),
             create_indent_trivia(ctx, shape),
         ]));
-        type_definition = hang_type_info(ctx, type_definition, shape);
+        type_definition = hang_type_info(ctx, type_declaration.type_definition(), shape);
     }
+
+    let type_definition =
+        type_definition.update_trailing_trivia(FormatTriviaType::Append(trailing_trivia));
 
     type_declaration
         .to_owned()