+++ tests/inputs-luau/type-hanging-comments.lua
@@ -1,0 +1,14 @@
+-- https://github.com/JohnnyMorganz/StyLua/issues/378
+export type KindEnum =
+	"Name" |
+	-- Document
+	"Document"
+	| "OperationDefinition"
+	| "VariableDefinition"
+	| "SelectionSet"
+	| "Field"
+	| "Argument" |
+	-- Fragments
+	"FragmentSpread"
+	| "InlineFragment"
+	| "FragmentDefinition"